% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twofold-Main.R
\docType{class}
\name{Twofold-Main-class}
\alias{Twofold-Main-class}
\alias{Twofold-Main}
\title{TwoFold-Main}
\description{
 An R package for data imputation
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula}

\item{\code{data}}{A data frame}
}}

\section{Methods}{

\describe{
\item{\code{TwoFCS()}}{This function which calls Twofold method.}

\item{\code{TwoFold()}}{This function that implements FCS-Twofold algorithm.}
}}
