% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ulr.R
\name{predict.ulr}
\alias{predict.ulr}
\title{Predict method for unary linear regression model fits}
\usage{
\method{predict}{ulr}(model, x)
}
\arguments{
\item{model, }{unary linear model}

\item{x, }{new data}
}
\value{
a numeric vector, the predicted values
}
\description{
Predicted values based on unary linear model object.
}
\details{
predict.ulr produces predicted values,
obtained by evaluating the regression function in the newdata x
}
