% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ulr.R
\name{print.ulr}
\alias{print.ulr}
\title{Print an unitary linear regression}
\usage{
\method{print}{ulr}(model)
}
\arguments{
\item{model, }{fitted ulr model}
}
\description{
print.ulr is to print an unitary linear regression model
}
\details{
Just like the famous print.lm,
print.ulr is to print the information of an unitary linear regression,
including the intercept and the slope.
}
