% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ulr.R
\name{summary.ulr}
\alias{summary.ulr}
\title{Summarizing unary linear regression model fits}
\usage{
\method{summary}{ulr}(model)
}
\arguments{
\item{model, }{unary linear model}
}
\value{
a list, more information about the model
}
\description{
summary method for class "ulr".
}
\details{
Summary for ulr model.
}
