% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulr.R
\name{ulr}
\alias{ulr}
\title{Fitting unitary linear regression model}
\usage{
ulr(x, y)
}
\arguments{
\item{x, }{independent variable}

\item{y, }{dependent variable}
}
\value{
ulr model
}
\description{
ulr is used to fit unitary linear regression model models
}
\details{
The model is fitted by OLS in the following way:
\deqn{B =(X^TX)^{-1}X^TY}
}
